+++ CHANGELOG.md
@@ -14,9 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Security
 -->
-## Unreleased
+## 1.0.40.8 - 2023-06-20
 ### Fixed
-- [RT#187856]: repeated reminders fail.
+- Automated repeated reminders failed
+
 ## 1.0.40.7 - 2023-06-08
 ### Fixed
 - [133](https://github.com/CCSDForge/episciences/issues/133): wrong acceptance date and docUrl for tmp versions.